feat: Neural Networks
Added pipeline for neural network in sknn. Changed parameters: Linear to Softmax. Implemented MinMaxScaler instead of StandardScaler to scale all data in the dataset to range (0, 1.0) for optimal performance in sknn Regressor module. Also changed the data type to np.ndarray as required for input type for Regressor algorithm in sknn module.